+++ include/orita.hpp
@@ -6,9 +6,10 @@
 #include"log.hpp"
 #include"files.hpp"
 #include"settings.hpp"
+#include"env.hpp"
+#include"process.hpp"
 #include"compile.hpp"
 #include"print.hpp"
 #include"judge.hpp"
 #include"app.hpp"
-#include"env.hpp"
 #endif