+++ CHANGELOG.md
@@ -279,3 +279,19 @@ Orita 第一个正式版本。
 - Use `Poco::Util`.
 - Use CMake to compile entrance file.
 - Use enviroment variables to pass arguments.
+
+### Dev 0.2.1
+
+#### Features
+
+- Multi-threaded compilation.
+- Multi-threaded check.
+- Multi-threaded judge.
+
+#### Fixes
+
+- Unexpected output when the result is `NULL`.
+
+#### Changes
+
+- Use `Poco::Process`.